+++ CHANGELOG.md
@@ -16,6 +16,12 @@
 **Note**: A feature tagged as Experimental is in a
 high state of flux, you're at risk of it changing without notice.
 
+# 2.11.10
+
+- **Polish**
+  - `TaskEither` / `TaskOption`
+    - now `tryCatch` / `tryCatchK` will catch synchronous errors too, #1676 (@thewilkybarkid)
+
 # 2.11.9
 
 - **Polish**